enh: separate in 2 modules
